backend/requirements.txt
# Backend requirements - SGT400 Predictive Maintenance API
# Reference: https://learn.microsoft.com/azure/app-service/quickstart-python

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0

# HTTP client
httpx==0.27.0

# Data processing
pandas==2.2.0
numpy==1.26.0
pyarrow==15.0.0

# ML / Inference
scikit-learn==1.4.0
scipy==1.12.0

# Azure SDKs
azure-identity==1.17.0
azure-keyvault-secrets==4.8.0
azure-ai-ml==1.19.0

# Monitoring
opencensus-ext-azure==1.1.13

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
httpx[http2]==0.27.0

# Linting
ruff==0.4.0
